cleanup of templates

+++ app/Http/Controllers/Admin/Dictionary/StackController.php
@@ -9,8 +9,13 @@ use App\Models\Dictionary\Stack;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Validation\Rule;
 use Illuminate\View\View;
+use Illuminate\Support\Str;
 
+/**
+ *
+ */
 class StackController extends BaseController
 {
     /**
@@ -23,8 +28,7 @@ class StackController extends BaseController
     {
         $perPage= $request->query('per_page', $this->perPage);
 
-        $stacks = Stack::orderBy('name', 'asc')
-            ->paginate($perPage);
+        $stacks = Stack::orderBy('name', 'asc')->paginate($perPage);
 
         return view('admin.dictionary.stack.index', compact('stacks'))
             ->with('i', (request()->input('page', 1) - 1) * $perPage);
@@ -32,25 +36,51 @@ class StackController extends BaseController
 
     /**
      * Show the form for creating a new stack.
+     *
+     * @param Request $request
+     * @return View
      */
     public function create(): View
     {
+        if (!Auth::guard('admin')->user()->root) {
+            abort(403, 'Only admins with root access can add stack entries.');
+        }
+
+        $referer = Request()->headers->get('referer');
+
         return view('admin.dictionary.stack.create');
     }
 
     /**
      * Store a newly created stack in storage.
+     *
+     * @param StackStoreRequest $request
+     * @return RedirectResponse
      */
     public function store(StackStoreRequest $request): RedirectResponse
     {
-        $stack =Stack::create($request->validated());
+        if (!Auth::guard('admin')->user()->root) {
+            abort(403, 'Only admins with root access can add stack entries.');
+        }
 
-        return redirect()->route('admin.dictionary.stack.show', $stack)
-            ->with('success', 'Stack created successfully.');
+        $stack = Stack::create($request->validated());
+
+        $referer = $request->input('referer');
+
+        if (!empty($referer)) {
+            return redirect(str_replace(config('app.url'), '', $referer))
+                ->with('success', $stack->name . ' created successfully.');
+        } else {
+            return redirect()->route('admin.dictionary.stack.show', $stack)
+                ->with('success', $stack->name . ' created successfully.');
+        }
     }
 
     /**
      * Display the specified stack.
+     *
+     * @param Stack $stack
+     * @return View
      */
     public function show(Stack $stack): View
     {
@@ -59,40 +89,75 @@ class StackController extends BaseController
 
     /**
      * Show the form for editing the specified stack.
+     *
+     * @param Stack $stack
+     * @param Request $request
+     * @return View
      */
-    public function edit(Stack $stack): View
+    public function edit(Stack $stack, Request $request): View
     {
         if (!Auth::guard('admin')->user()->root) {
             abort(403, 'Only admins with root access can edit stack entries.');
         }
 
-        return view('admin.dictionary.stack.edit', compact('stack'));
+        $referer = $request->headers->get('referer');
+
+        return view('admin.dictionary.stack.edit', compact('stack', 'referer'));
     }
 
     /**
      * Update the specified stack in storage.
+     *
+     * @param StackUpdateRequest $request
+     * @param Stack $stack
+     * @return RedirectResponse
      */
-    public function update(StackUpdateRequest $request,
-                           Stack              $stack): RedirectResponse
+    public function update(StackUpdateRequest $request, Stack $stack): RedirectResponse
     {
         if (!Auth::guard('admin')->user()->root) {
             abort(403, 'Only admins with root access can update stack entries.');
         }
 
+        // Validate the posted data and generated slug.
+        $validatedData = $request->validated();
+        $request->merge([ 'slug' => Str::slug($validatedData['name']) ]);
+        $request->validate(['slug' => [ Rule::unique('posts', 'slug') ] ]);
         $stack->update($request->validated());
 
-        return redirect()->route('admin.dictionary.stack.show', $stack)
-            ->with('success', 'Stack updated successfully');
+        $referer = $request->input('referer');
+
+        if (!empty($referer)) {
+            return redirect(str_replace(config('app.url'), '', $referer))
+                ->with('success', $stack->name . ' updated successfully.');
+        } else {
+            return redirect()->route('admin.dictionary.stack.show', $stack)
+                ->with('success', $stack->name . ' updated successfully');
+        }
     }
 
     /**
      * Remove the specified stack from storage.
+     *
+     * @param Stack $stack
+     * @param Request $request
+     * @return RedirectResponse
      */
-    public function destroy(Stack $stack): RedirectResponse
+    public function destroy(Stack $stack, Request $request): RedirectResponse
     {
+        if (!Auth::guard('admin')->user()->root) {
+            abort(403, 'Only admins with root access can delete stack entries.');
+        }
+
         $stack->delete();
 
-        return redirect()->route('admin.dictionary.stack.index')
-            ->with('success', 'Stack deleted successfully');
+        $referer = $request->input('referer');
+
+        if (!empty($referer)) {
+            return redirect(str_replace(config('app.url'), '', $referer))
+                ->with('success', $stack->name . ' deleted successfully.');
+        } else {
+            return redirect()->route('admin.dictionary.stack.index')
+                ->with('success', $stack->name . ' deleted successfully');
+        }
     }
 }
